+++ src/main/java/gregicadditions/integrations/bees/ForestryCommonProxy.java
@@ -51,8 +51,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static gregicadditions.machines.GATileEntities.location;
-import static gregicadditions.recipes.helper.AdditionMethods.registerMachineRecipe;
-import static gregicadditions.recipes.helper.AdditionMethods.removeRecipesByInputs;
+import static gregicadditions.recipes.helper.HelperMethods.registerMachineRecipe;
+import static gregicadditions.recipes.helper.HelperMethods.removeRecipesByInputs;
 import static gregicadditions.recipes.helper.GACraftingComponents.*;
 
 @Mod.EventBusSubscriber()
